fix macos tests
